Split single test file into multiple ones
